Some changes to data mdoel to start it all off
