+++ engine_datamodel/include/data_model_config.hpp
@@ -1,0 +1,9 @@
+#ifndef RCBE_DATA_MODEL_CONFIG_HPP
+#define RCBE_DATA_MODEL_CONFIG_HPP
+
+namespace rcbe::data_model
+{
+using EngineScalar = double;
+}
+
+#endif
