+++ engine_datamodel/include/data_types/rendering_types.hpp
@@ -2,6 +2,7 @@
 #define RCBE_DATA_TYPES_RENDERING_TYPES
 
 #include <data_types/core_types.hpp>
+#include <geometry/geometry.hpp>
 
 namespace rcbe::data_types::rendering
 {
@@ -12,17 +13,17 @@ class RenderingObject
 public:
     virtual void render() = 0;
 
-    void SetTranslation(core::Vector3d &&translation);
-    void SetOrientation(core::Matrix3x3d &&orientation);
+    void SetTranslation(geometry::Vector3d &&translation);
+    void SetOrientation(geometry::Matrix3x3d &&orientation);
     void setId(std::string &id);
 
-    const core::Vector3d &GetTranslation() const;
-    const core::Matrix3x3d &GetOrientation() const;
+    const geometry::Vector3d &GetTranslation() const;
+    const geometry::Matrix3x3d &GetOrientation() const;
     const std::string &GetId() const;
 
 private:
-    core::Vector3d _translation;
-    core::Matrix3x3d _orientation;
+    geometry::Vector3d _translation;
+    geometry::Matrix3x3d _orientation;
 
     std::string _id;
     std::vector<std::string> _tags;
